fix: fetch-google-repo-tool fetcher
Prior to this bug, the fetcher would inadvertently update the manifest
itself, which would every once in a while cause havoc. I added a,
mitigation a couple of years back, which mostly worked --- but only
mostly. Also, it was a terrible hack.

This commit removes the mitigation and fixes the root-cause, via the
power of the `--no-manifest-update` flag.

Signed-off-by: wucke13 <wucke13+github@gmail.com>